pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backup-projects"
version = "0.1.0"
description = "Backup projects v1 foundation"
requires-python = ">=3.12"
license = { file = "LICENSE" }
dependencies = [
  "Flask>=3.0,<4.0",
  "Jinja2>=3.1,<4.0",
  "Typer>=0.12,<1.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.4,<1.0",
  "black>=24.0,<26.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
